WA-275 Transactions actions refactoring
