+++ app/actions/transactions/addTransaction.js
@@ -1,0 +1,7 @@
+export const ADD_TRANSACTION = 'ADD_TRANSACTION'
+
+export const addTransaction = (tx, popupId) => ({
+  type: ADD_TRANSACTION,
+  tx,
+  popupId
+})
