+++ app/actions/transactions/removeAllTransactions.js
@@ -1,0 +1,5 @@
+export const REMOVE_ALL_TRANSACTIONS = 'REMOVE_ALL_TRANSACTIONS'
+
+export const removeAllTransactions = () => ({
+  type: REMOVE_ALL_TRANSACTIONS,
+})
