+++ app/actions/transactions/removeTransaction.js
@@ -1,0 +1,6 @@
+export const REMOVE_TRANSACTION = 'REMOVE_TRANSACTION'
+
+export const removeTransaction = (popupId) => ({
+  type: REMOVE_TRANSACTION,
+  popupId
+})
